refactor the code in main to avoid require it in the test file
